+++ src/Compilers/CSharp/Portable/Compilation/CSharpCompilation.cs
@@ -4079,7 +4079,7 @@ namespace Microsoft.CodeAnalysis.CSharp
             }
         }
 
-        internal override DeterministicKeyBuilder CreateDetermisiticKeyBuilder() => new CSharpDeterministicKeyBuilder();
+        internal override DeterministicKeyBuilder CreateDeterministicKeyBuilder() => new CSharpDeterministicKeyBuilder();
 
         private ImmutableArray<string> GetPreprocessorSymbols()
         {
